Title: 합과 차의 나눗셈, Time: 8ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240123/합과 차의 나눗셈/divide-of-sum-and-sub.c
@@ -2,10 +2,10 @@
 
 int main() {
     // 여기에 코드를 작성해주세요.
-    int a,b;
-    scanf("%d %d",&a,&b);
+    float a,b;
+    scanf("%f %f",&a,&b);
 
     printf("%.2f",(a+b)/(a-b));
-    
+
     return 0;
 }
